Added API and Model Interface

+++ FSA.Domain/Entities/FSAClaim.cs
@@ -9,11 +9,11 @@ namespace FSA.Domain.Entities
     public class FSAClaim
     {
         public int ID { get; set; }
-        public int ClaimAmount { get; set; }
+        public decimal ClaimAmount { get; set; }
         public DateTime DateSubmitted { get; set; }
         public string ReceiptNumber { get; set; }
         public DateTime ReceiptDate { get; set; }
-        public int ReceiptAmount { get; set; }
+        public decimal ReceiptAmount { get; set; }
         public string Status { get; set; } = "Pending";
         public int EmployeeID { get; set; }
         public DateTime Modified { get; set; }
